requirements.txt
langchain
langchain-google-genai
langchain-community
pypdf
faiss-cpu
python-dotenv
flask